requirements.txt
Django~=3.2.10
django-ckeditor
django-model-utils
mysqlclient
Pillow